Filter på produktside (#440)
* Endre min/max filter til å håndtere enten eller situasjoner

* Hjelpetekst på filteroverskrifteb

* Hent filter i eget kall

* Filtrering funker, verdien i formet er ikke oppdatert

* Legg til filter state på defaultvalkues

* Add chips, men de er treige :thinking-face:

* Sett min på inputfeltene for å unngå negative tall

* Gjør det mulig å fjerne søkeord inne på produktsiden

* Lenker på produktkort tar med params

* Send med filter til produktsiden men ikke andre vei

* Url oppdateres etter tilgjengelige filter

* Styling på filter som åpnes over innhold under

* deaktiver vis filter på produktside enn så lenge

* Legg search term til filterChips object og vis om det er HMS eller lev-art

* Flytt litt ut i en utils-fil

+++ src/utils/sort-util.ts
@@ -6,7 +6,7 @@ const stringComparator = (a: string, b: string) => {
   if (b === '-' && a !== '-') {
     return -1 // "-" should come last
   }
-  
+
   if (a < b || b === undefined) {
     return -1
   }
@@ -21,10 +21,10 @@ export const sortAlphabetically = (keyA: string, keyB: string, desc: boolean = f
 }
 
 export const sortIntWithStringFallback = (keyA: string, keyB: string, desc: boolean = false) => {
-  if (parseInt(keyA) && parseInt(keyB)) {
-    return desc ? parseInt(keyB) - parseInt(keyA) : parseInt(keyA) - parseInt(keyB)
+  if (parseFloat(keyA) && parseFloat(keyB)) {
+    return desc ? parseFloat(keyB) - parseFloat(keyA) : parseFloat(keyA) - parseFloat(keyB)
   }
-  
+
   return desc ? stringComparator(keyB, keyA) : stringComparator(keyA, keyB)
 }
